fixed issue where stream settings were not inited
